Address Form ++
Made several changes to the site:

- Added the rough draft addressing form, its corresponding PHP backend,
and changed the Addressing page to link to the address request form
- Added a tentative link to the online job application on the Contact
page, as well as uploaded the application form pdf and placed a link
underneath the online application for it
- Altered the index page so that it will adjust better to different
resolutions.
-  Made the Smart 911 sidebar a bar that goes across the top of the
index page, giving it more prominence as well as allowing it to adjust
to resolutions better.
- Replaced the Facebook Video link with a YouTube link of the same
video. After changing the smart911 banner, the facebook video was taking
up the whole page, which is really large on HD or better resolutions.
YouTube video link allowed me to resize and also place it on the screen
better.