Address Request Form Update + Others
Added new address request form with javascript, also added pictures that
were sent to the website. Four were added to the Smart 911 page, and
most of the rest were added to the carousel that is on the main page. A
few pictures were not added, mainly due to picture quality.