reCAPTCHA added
Added Captcha code, made some changes to the display page, and changed
the emailer back over to PEAR so that we don't have to mess with any
php.ini files. Also reorganized everything so that it flowed better, and
also made it where there wasn't any possible overlapping javascript or
css files.